refactor(api): extract extractPgCode to lib/pg-error.ts
Removes the verbatim duplicate in users.ts and error-handler.ts,
consolidating into a single shared helper.
